Fix issue where relative path needs to be converted to absolute path
This problem is caused because the copy_files subroutine tries to copy files
*while* scanning for files. PWD of the program changes while it is scanning
through directories, thus, copy() can't find the files and directories if
relative paths are used. As a solution, copy_file() is renamed to
add_files_to_copy which adds the partial path under $from to copy (so that the
destination path can be easily created by concatenating the $to) and later
using the copy_files() to copy the files to the destination.