Add overload constructor without redundant parameters
This change gives rise to the issue regarding the texture. Since the Texture
in LevelScreen class is only used as an argument, instantiation of the Texture
is done in the Car class. However, the Texture in the LevelScreen class is not
removed in case it is planned to use that variable in future commits. A TODO
comment addressing this suggestion is added.