Temporarily change the parser to a public variable
This change is made because issue arises when instantiating two MapParser
objects (i.e. going over the initialization of the maps and traffic signals
twice) where cars stop at the first traffic signal they encounter. This issue
is probably caused by the traffic signals not being hooked up to the grids.
Thus, in order to make the reading of the XML file only once, the parser is
made available to the LevelScreen class by making the visibility public.