Added button generation methods for MainMenu screen
Added both text and image button generation methods for user
interactivity on the screen; as well as step by step comments on how to
place and render them. level1 and level2 are just cheap images I threw
together in paint and should be replaced. title2 is just a better
looking title screen, which too could be replaced.