Add local server and remote server
This commit adds a local server and a remote server. The local server receives
data such as results of levels and time spent in the game and editor, and
sends them to the remote server. The remote server currently only prints out
the received data to stdout. The editor does not send the time spent in the
editor to the local server yet.